Position, PositionDao, TestPositionDao
buat Entity Position
buat Dao PositionDao
buat JUnit TestPositionDao
