+++ rest-web/src/main/resources/gdn-project-context.xml
@@ -8,8 +8,8 @@
         http://www.springframework.org/schema/task http://www.springframework.org/schema/task/spring-task.xsd
 		http://www.ryantenney.com/schema/metrics http://www.ryantenney.com/schema/metrics/metrics-3.0.xsd">
   <import resource="classpath:gdn-jpa-configuration.xml" />
-  <jpa:repositories base-package="com.ega.dao" />
-  <context:component-scan base-package="com.ega.services" />
+  <jpa:repositories base-package="com.gdn.x.beirut.dao" />
+  <context:component-scan base-package="com.gdn.x.beirut.services" /> 
   <!-- <context:component-scan base-package="com.gdn.x.productcategorybase.config" /> -->
   <context:property-placeholder location="file:///D:/Kuliah/Prosi/Project_Config/beirut/beirut.properties"
     properties-ref="applicationDefaultProperties" ignore-resource-not-found="true" ignore-unresolvable="true" />
@@ -21,7 +21,7 @@
 <!--     <constructor-arg ref="dataSource" /> -->
 <!--   </bean> -->
   <bean id="dozerMapper" class ="org.dozer.DozerBeanMapper"/>
-  
+
   <task:executor id="taskExecutor" queue-capacity="10" pool-size="${thread.pool.size}" rejection-policy="CALLER_RUNS" />
   <bean id="executorService" class="org.springframework.core.task.support.ExecutorServiceAdapter">
     <constructor-arg ref="taskExecutor" />
