+++ rest-web/src/main/resources/gdn-x-controller.xml
@@ -11,6 +11,6 @@
         http://www.springframework.org/schema/task http://www.springframework.org/schema/task/spring-task.xsd
         http://www.springframework.org/schema/security http://www.springframework.org/schema/security/spring-security.xsd">
          
-    <context:component-scan base-package="com.ega.controller" />
+    <context:component-scan base-package="com.gdn.x.beirut.controller" />
      
 </beans>
